RESOLVED - issue BATCH-194: Incorrect exception handling when using Hibernate http://opensource.atlassian.com/projects/spring/browse/BATCH-194
Created HibernateAwareItemWriter to abstract the flushing concerns into a framework class - to use it you have to register it with the chunkOperations as an interceptor *and* inject it with an ItemWriter for the step concerned.  There might be an AOP approach to this that would reduce the burden of having to confugure it - we can look into that as part of the namespace / DSL work.  See hibernateJob.xml for sample.