RESOLVED - issue BATCH-199: ExceptionHandler in SImpleStepConfiguration ignored by SimpleStepConfigurationFactory http://opensource.atlassian.com/projects/spring/browse/BATCH-199
Moved ExecptionHandler to the step operations.

It seemed to me that the responsibility for doing this should be encapsulated in the StepExecutor - also prevents a nasty package cycle.  The only way to do this and keep everyone honest was to add an applyConfiguration() method to the StepExecutor interface.  I am open to suggestion about how to do this more cleanly, but for now that is probably enough.