Further finalizing public API.  A couple of big changes:
1) Exracted all enhancement-related artifacts out of .support and into .enhancement
2) Eliminated the final cyclic dependency by consolidating existing listeners into the .process package (DefaultConfigurationListenerRegistry was causing the cycle, because it dug down into the .listener.* packgaes).  These listener implementations have also been made package-private, and the ConfigurationListenerRegistry interface has been eliminated in favor of a base class by the same name (containing the code that DefaultConfigurationListenerRegistry used to).

The rationale for (2) above is that the listener implementations are *required* for proper internal functioning.  They are not designed for public consumption.  Likewise, the ConfigurationListenerRegistry interface gave users the illusion that they could directly implement this interface and be good to go.  This is not so; it is is required that all of the internal ConfigurationListeners be registered.  So by eliminating the interface and moving all default listener registration into in aninstance initializer in the base class, we absolutely guarantee registration of required listeners *and* allow users to subclass and extend.  By moving all internal listener implementations local to the .process package, we eliminate the cyclic dependency problem, and by making them package-private, we reduce the surface area of the public API.

All other modifications in this commit are generally in support of the above two changes.